테스트중

+++ src/main/java/com/codehows/taelimbe/ai/service/QnaService.java
@@ -1,12 +1,14 @@
 package com.codehows.taelimbe.ai.service;
 
 import com.codehows.taelimbe.ai.constant.QnaStatus;
+import com.codehows.taelimbe.ai.entity.Embed;
 import com.codehows.taelimbe.ai.entity.Qna;
 import com.codehows.taelimbe.ai.repository.EmbedRepository;
 import com.codehows.taelimbe.ai.repository.QnaRepository;
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -17,7 +19,7 @@ public class QnaService {
     private final QnaRepository qnaRepository;
     private final EmbeddingService embeddingService;
     private final EmbedRepository embedRepository;
-
+    private final QnaEmbeddingFailService qnaEmbeddingFailService;
 
 
 
@@ -33,27 +35,30 @@ public class QnaService {
             qna.applySuccess();
 
         } catch (Exception e) {
-            qna.applyFail();
+            qnaEmbeddingFailService.markFail(qnaId, answer);
             throw e;
         }
     }
-
-
-
-
 
 
     @Transactional
     public void questionDelete(Long qnaId) {
         Qna qna = get(qnaId);
 
-        embedRepository.findByQnaId(qnaId)
-                .forEach(embed ->
-                        embeddingService.deleteByKey(embed.getEmbedKey())
-                );
+        // Embed 조회
+        Embed embed = embedRepository.findByQnaId(qnaId)
+                .orElse(null);
 
+        // Milvus + RDB 삭제
+        if (embed != null) {
+            embeddingService.deleteByKey(embed.getEmbedKey());
+            embedRepository.delete(embed);
+        }
+
+        // QnA 삭제
         qnaRepository.delete(qna);
     }
+
 
     public Qna get(Long qnaId) {
         return qnaRepository.findById(qnaId)
